dataloading/oxo/requirements.txt
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
idna==2.6
mysqlclient==1.3.12
neo4j-driver==1.5.3
PyYAML==3.12
requests==2.18.4
urllib3==1.22
